Port https://github.com/redis-rs/redis-rs/pull/1711 to 0.32.x
This is based on the work done in the valkey-glide project for mapping commands and their routing info.
