+++ redis/src/cluster_handling/async_connection/routing.rs
@@ -1,7 +1,7 @@
 use crate::{
     cluster_routing::{
-        self, MultipleNodeRoutingInfo, Redirect, ResponsePolicy, Route, SingleNodeRoutingInfo,
-        SlotAddr,
+        MultipleNodeRoutingInfo, Redirect, ResponsePolicy, Route, RoutingInfo,
+        SingleNodeRoutingInfo, SlotAddr,
     },
     Cmd, ErrorKind, RedisResult,
 };
@@ -12,15 +12,11 @@ pub(super) enum InternalRoutingInfo<C> {
     MultiNode((MultipleNodeRoutingInfo, Option<ResponsePolicy>)),
 }
 
-impl<C> From<cluster_routing::RoutingInfo> for InternalRoutingInfo<C> {
-    fn from(value: cluster_routing::RoutingInfo) -> Self {
+impl<C> From<RoutingInfo> for InternalRoutingInfo<C> {
+    fn from(value: RoutingInfo) -> Self {
         match value {
-            cluster_routing::RoutingInfo::SingleNode(route) => {
-                InternalRoutingInfo::SingleNode(route.into())
-            }
-            cluster_routing::RoutingInfo::MultiNode(routes) => {
-                InternalRoutingInfo::MultiNode(routes)
-            }
+            RoutingInfo::SingleNode(route) => InternalRoutingInfo::SingleNode(route.into()),
+            RoutingInfo::MultiNode(routes) => InternalRoutingInfo::MultiNode(routes),
         }
     }
 }
@@ -62,21 +58,25 @@ impl<C> From<SingleNodeRoutingInfo> for InternalSingleNodeRouting<C> {
             SingleNodeRoutingInfo::ByAddress { host, port } => {
                 InternalSingleNodeRouting::ByAddress(format!("{host}:{port}"))
             }
+            SingleNodeRoutingInfo::RandomPrimary => {
+                InternalSingleNodeRouting::SpecificNode(Route::new_random_primary())
+            }
         }
     }
 }
 
 pub(super) fn route_for_pipeline(pipeline: &crate::Pipeline) -> RedisResult<Option<Route>> {
     fn route_for_command(cmd: &Cmd) -> Option<Route> {
-        match cluster_routing::RoutingInfo::for_routable(cmd) {
-            Some(cluster_routing::RoutingInfo::SingleNode(SingleNodeRoutingInfo::Random)) => None,
-            Some(cluster_routing::RoutingInfo::SingleNode(
-                SingleNodeRoutingInfo::SpecificNode(route),
-            )) => Some(route),
-            Some(cluster_routing::RoutingInfo::MultiNode(_)) => None,
-            Some(cluster_routing::RoutingInfo::SingleNode(SingleNodeRoutingInfo::ByAddress {
-                ..
-            })) => None,
+        match RoutingInfo::for_routable(cmd) {
+            Some(RoutingInfo::SingleNode(SingleNodeRoutingInfo::Random)) => None,
+            Some(RoutingInfo::SingleNode(SingleNodeRoutingInfo::SpecificNode(route))) => {
+                Some(route)
+            }
+            Some(RoutingInfo::MultiNode(_)) => None,
+            Some(RoutingInfo::SingleNode(SingleNodeRoutingInfo::ByAddress { .. })) => None,
+            Some(RoutingInfo::SingleNode(SingleNodeRoutingInfo::RandomPrimary)) => {
+                Some(Route::new_random_primary())
+            }
             None => None,
         }
     }
@@ -91,7 +91,7 @@ pub(super) fn route_for_pipeline(pipeline: &crate::Pipeline) -> RedisResult<Opti
             (Some(chosen_route), Some(next_cmd_route)) => {
                 if chosen_route.slot() != next_cmd_route.slot() {
                     Err((ErrorKind::CrossSlot, "Received crossed slots in pipeline").into())
-                } else if chosen_route.slot_addr() != &SlotAddr::Master {
+                } else if chosen_route.slot_addr() != SlotAddr::Master {
                     Ok(Some(next_cmd_route))
                 } else {
                     Ok(Some(chosen_route))
