+++ redis/src/cluster_handling/sync_connection/mod.rs
@@ -64,33 +64,32 @@
 //! ```
 use std::cell::RefCell;
 use std::collections::HashSet;
-use std::str::FromStr;
 use std::thread;
 use std::time::Duration;
 
-use crate::cluster_pipeline::UNROUTABLE_ERROR;
+mod pipeline;
+
+pub use super::client::{ClusterClient, ClusterClientBuilder};
+use super::topology::parse_slots;
+use super::{
+    client::ClusterParams,
+    routing::{Redirect, Route, RoutingInfo},
+    slot_map::{SlotMap, SLOT_SIZE},
+};
+use crate::cluster_handling::{get_connection_info, slot_cmd, split_node_address};
 use crate::cluster_routing::{
     MultipleNodeRoutingInfo, ResponsePolicy, Routable, SingleNodeRoutingInfo, SlotAddr,
 };
-use crate::cluster_topology::parse_slots;
 use crate::cmd::{cmd, Cmd};
-use crate::connection::{
-    connect, Connection, ConnectionAddr, ConnectionInfo, ConnectionLike, RedisConnectionInfo,
-};
+use crate::connection::{connect, Connection, ConnectionInfo, ConnectionLike};
 use crate::parser::parse_redis_value;
 use crate::types::{ErrorKind, HashMap, RedisError, RedisResult, Value};
 use crate::IntoConnectionInfo;
 pub use crate::TlsMode; // Pub for backwards compatibility
-use crate::{
-    cluster_client::ClusterParams,
-    cluster_routing::{Redirect, Route, RoutingInfo, SlotMap, SLOT_SIZE},
-};
+use pipeline::UNROUTABLE_ERROR;
 use rand::{rng, seq::IteratorRandom, Rng};
 
-pub use crate::cluster_client::{ClusterClient, ClusterClientBuilder};
-pub use crate::cluster_pipeline::{cluster_pipe, ClusterPipeline};
-
-use crate::connection::TlsConnParams;
+pub use pipeline::{cluster_pipe, ClusterPipeline};
 
 #[derive(Clone)]
 enum Input<'a> {
@@ -688,7 +687,7 @@ where
         let mut slots = self.slots.borrow_mut();
 
         let results = match &routing {
-            MultipleNodeRoutingInfo::MultiSlot(routes) => {
+            MultipleNodeRoutingInfo::MultiSlot((routes, _)) => {
                 self.execute_multi_slot(input, &mut slots, &mut connections, routes)
             }
             MultipleNodeRoutingInfo::AllMasters => {
@@ -731,18 +730,35 @@ where
                 Err(last_failure
                     .unwrap_or_else(|| (ErrorKind::IoError, "Couldn't find a connection").into()))
             }
-            Some(ResponsePolicy::OneSucceededNonEmpty) => {
+            Some(ResponsePolicy::CombineMaps) => crate::cluster_routing::combine_map_results(
+                results
+                    .into_iter()
+                    .map(|result| result.map(|(_, value)| value))
+                    .collect::<RedisResult<Vec<_>>>()?,
+            ),
+            Some(ResponsePolicy::FirstSucceededNonEmptyOrAllEmpty) => {
+                // Attempt to return the first result that isn't `Nil` or an error.
+                // If no such response is found and all servers returned `Nil`, it indicates that all shards are empty, so return `Nil`.
+                // If we received only errors, return the last received error.
+                // If we received a mix of errors and `Nil`s, we can't determine if all shards are empty,
+                // thus we return the last received error instead of `Nil`.
                 let mut last_failure = None;
-
+                let num_of_results = results.len();
+                let mut nil_counter = 0;
                 for result in results {
                     match result.map(|(_, res)| res) {
-                        Ok(Value::Nil) => continue,
+                        Ok(Value::Nil) => nil_counter += 1,
                         Ok(val) => return Ok(val),
                         Err(err) => last_failure = Some(err),
                     }
                 }
-                Err(last_failure
-                    .unwrap_or_else(|| (ErrorKind::IoError, "Couldn't find a connection").into()))
+                if nil_counter == num_of_results {
+                    Ok(Value::Nil)
+                } else {
+                    Err(last_failure.unwrap_or_else(|| {
+                        (ErrorKind::IoError, "Couldn't find a connection").into()
+                    }))
+                }
             }
             Some(ResponsePolicy::Aggregate(op)) => {
                 let results = results
@@ -764,10 +780,9 @@ where
                     .map(|res| res.map(|(_, val)| val))
                     .collect::<RedisResult<Vec<_>>>()?;
                 match routing {
-                    MultipleNodeRoutingInfo::MultiSlot(vec) => {
+                    MultipleNodeRoutingInfo::MultiSlot((vec, pattern)) => {
                         crate::cluster_routing::combine_and_sort_array_results(
-                            results,
-                            vec.iter().map(|(_, indices)| indices),
+                            results, &vec, &pattern,
                         )
                     }
                     _ => crate::cluster_routing::combine_array_results(results),
@@ -835,6 +850,9 @@ where
                             let address = format!("{host}:{port}");
                             let conn = self.get_connection_by_addr(&mut connections, &address);
                             (address, conn)
+                        }
+                        SingleNodeRoutingInfo::RandomPrimary => {
+                            self.get_connection(&mut connections, &Route::new_random_primary())
                         }
                     }
                 };
@@ -1113,70 +1131,10 @@ fn get_random_connection_or_error<C: ConnectionLike + Connect + Sized>(
     }
 }
 
-// The node string passed to this function will always be in the format host:port as it is either:
-// - Created by calling ConnectionAddr::to_string (unix connections are not supported in cluster mode)
-// - Returned from redis via the ASK/MOVED response
-pub(crate) fn get_connection_info(
-    node: &str,
-    cluster_params: ClusterParams,
-) -> RedisResult<ConnectionInfo> {
-    let (host, port) = split_node_address(node)?;
-
-    Ok(ConnectionInfo {
-        addr: get_connection_addr(host, port, cluster_params.tls, cluster_params.tls_params),
-        redis: RedisConnectionInfo {
-            password: cluster_params.password,
-            username: cluster_params.username,
-            protocol: cluster_params.protocol.unwrap_or_default(),
-            db: 0,
-        },
-    })
-}
-
-fn split_node_address(node: &str) -> RedisResult<(String, u16)> {
-    let invalid_error =
-        || RedisError::from((ErrorKind::InvalidClientConfig, "Invalid node string"));
-    node.rsplit_once(':')
-        .and_then(|(host, port)| {
-            Some(host.trim_start_matches('[').trim_end_matches(']'))
-                .filter(|h| !h.is_empty())
-                .map(str::to_string)
-                .zip(u16::from_str(port).ok())
-        })
-        .ok_or_else(invalid_error)
-}
-
-pub(crate) fn get_connection_addr(
-    host: String,
-    port: u16,
-    tls: Option<TlsMode>,
-    tls_params: Option<TlsConnParams>,
-) -> ConnectionAddr {
-    match tls {
-        Some(TlsMode::Secure) => ConnectionAddr::TcpTls {
-            host,
-            port,
-            insecure: false,
-            tls_params,
-        },
-        Some(TlsMode::Insecure) => ConnectionAddr::TcpTls {
-            host,
-            port,
-            insecure: true,
-            tls_params,
-        },
-        _ => ConnectionAddr::Tcp(host, port),
-    }
-}
-
-pub(crate) fn slot_cmd() -> Cmd {
-    let mut cmd = Cmd::new();
-    cmd.arg("CLUSTER").arg("SLOTS");
-    cmd
-}
-
 #[cfg(test)]
 mod tests {
+    use crate::ConnectionAddr;
+
     use super::*;
 
     #[test]
