+++ redis/src/commands/mod.rs
@@ -29,7 +29,7 @@ pub use json::JsonCommands;
 pub use json::JsonAsyncCommands;
 
 #[cfg(feature = "cluster")]
-use crate::cluster_pipeline::ClusterPipeline;
+use crate::cluster_handling::sync_connection::ClusterPipeline;
 
 #[cfg(feature = "geospatial")]
 use crate::geo;
@@ -45,16 +45,92 @@ use crate::RedisConnectionInfo;
 pub(crate) fn is_readonly_cmd(cmd: &[u8]) -> bool {
     matches!(
         cmd,
-        b"BITCOUNT"
+        b"ACL CAT"
+            | b"ACL DELUSER"
+            | b"ACL DRYRUN"
+            | b"ACL GENPASS"
+            | b"ACL GETUSER"
+            | b"ACL HELP"
+            | b"ACL LIST"
+            | b"ACL LOAD"
+            | b"ACL LOG"
+            | b"ACL SAVE"
+            | b"ACL SETUSER"
+            | b"ACL USERS"
+            | b"ACL WHOAMI"
+            | b"AUTH"
+            | b"BGREWRITEAOF"
+            | b"BGSAVE"
+            | b"BITCOUNT"
             | b"BITFIELD_RO"
             | b"BITPOS"
+            | b"CLIENT ID"
+            | b"CLIENT CACHING"
+            | b"CLIENT CAPA"
+            | b"CLIENT GETNAME"
+            | b"CLIENT GETREDIR"
+            | b"CLIENT HELP"
+            | b"CLIENT INFO"
+            | b"CLIENT KILL"
+            | b"CLIENT LIST"
+            | b"CLIENT NO-EVICT"
+            | b"CLIENT NO-TOUCH"
+            | b"CLIENT PAUSE"
+            | b"CLIENT REPLY"
+            | b"CLIENT SETINFO"
+            | b"CLIENT SETNAME"
+            | b"CLIENT TRACKING"
+            | b"CLIENT TRACKINGINFO"
+            | b"CLIENT UNBLOCK"
+            | b"CLIENT UNPAUSE"
+            | b"CLUSTER COUNT-FAILURE-REPORTS"
+            | b"CLUSTER COUNTKEYSINSLOT"
+            | b"CLUSTER FAILOVER"
+            | b"CLUSTER GETKEYSINSLOT"
+            | b"CLUSTER HELP"
+            | b"CLUSTER INFO"
+            | b"CLUSTER KEYSLOT"
+            | b"CLUSTER LINKS"
+            | b"CLUSTER MYID"
+            | b"CLUSTER MYSHARDID"
+            | b"CLUSTER NODES"
+            | b"CLUSTER REPLICATE"
+            | b"CLUSTER SAVECONFIG"
+            | b"CLUSTER SHARDS"
+            | b"CLUSTER SLOTS"
+            | b"COMMAND COUNT"
+            | b"COMMAND DOCS"
+            | b"COMMAND GETKEYS"
+            | b"COMMAND GETKEYSANDFLAGS"
+            | b"COMMAND HELP"
+            | b"COMMAND INFO"
+            | b"COMMAND LIST"
+            | b"CONFIG GET"
+            | b"CONFIG HELP"
+            | b"CONFIG RESETSTAT"
+            | b"CONFIG REWRITE"
+            | b"CONFIG SET"
             | b"DBSIZE"
             | b"DUMP"
-            | b"EVALSHA_RO"
+            | b"ECHO"
             | b"EVAL_RO"
+            | b"EVALSHA_RO"
             | b"EXISTS"
             | b"EXPIRETIME"
             | b"FCALL_RO"
+            | b"FT.AGGREGATE"
+            | b"FT.EXPLAIN"
+            | b"FT.EXPLAINCLI"
+            | b"FT.INFO"
+            | b"FT.PROFILE"
+            | b"FT.SEARCH"
+            | b"FT._ALIASLIST"
+            | b"FT._LIST"
+            | b"FUNCTION DUMP"
+            | b"FUNCTION HELP"
+            | b"FUNCTION KILL"
+            | b"FUNCTION LIST"
+            | b"FUNCTION STATS"
             | b"GEODIST"
             | b"GEOHASH"
             | b"GEOPOS"
@@ -64,56 +140,114 @@ pub(crate) fn is_readonly_cmd(cmd: &[u8]) -> bool {
             | b"GET"
             | b"GETBIT"
             | b"GETRANGE"
+            | b"HELLO"
             | b"HEXISTS"
-            | b"HEXPIRETIME"
             | b"HGET"
             | b"HGETALL"
             | b"HKEYS"
             | b"HLEN"
             | b"HMGET"
             | b"HRANDFIELD"
-            | b"HPTTL"
-            | b"HPEXPIRETIME"
             | b"HSCAN"
             | b"HSTRLEN"
-            | b"HTTL"
             | b"HVALS"
+            | b"JSON.ARRINDEX"
+            | b"JSON.ARRLEN"
+            | b"JSON.DEBUG"
+            | b"JSON.GET"
+            | b"JSON.OBJLEN"
+            | b"JSON.OBJKEYS"
+            | b"JSON.MGET"
+            | b"JSON.RESP"
+            | b"JSON.STRLEN"
+            | b"JSON.TYPE"
+            | b"INFO"
             | b"KEYS"
+            | b"LASTSAVE"
+            | b"LATENCY DOCTOR"
+            | b"LATENCY GRAPH"
+            | b"LATENCY HELP"
+            | b"LATENCY HISTOGRAM"
+            | b"LATENCY HISTORY"
+            | b"LATENCY LATEST"
+            | b"LATENCY RESET"
             | b"LCS"
             | b"LINDEX"
             | b"LLEN"
             | b"LOLWUT"
             | b"LPOS"
             | b"LRANGE"
+            | b"MEMORY DOCTOR"
+            | b"MEMORY HELP"
+            | b"MEMORY MALLOC-STATS"
+            | b"MEMORY PURGE"
+            | b"MEMORY STATS"
             | b"MEMORY USAGE"
             | b"MGET"
+            | b"MODULE HELP"
+            | b"MODULE LIST"
+            | b"MODULE LOAD"
+            | b"MODULE LOADEX"
+            | b"MODULE UNLOAD"
             | b"OBJECT ENCODING"
             | b"OBJECT FREQ"
+            | b"OBJECT HELP"
             | b"OBJECT IDLETIME"
             | b"OBJECT REFCOUNT"
             | b"PEXPIRETIME"
             | b"PFCOUNT"
+            | b"PING"
             | b"PTTL"
+            | b"PUBLISH"
+            | b"PUBSUB CHANNELS"
+            | b"PUBSUB HELP"
+            | b"PUBSUB NUMPAT"
+            | b"PUBSUB NUMSUB"
+            | b"PUBSUB SHARDCHANNELS"
+            | b"PUBSUB SHARDNUMSUB"
             | b"RANDOMKEY"
+            | b"REPLICAOF"
+            | b"RESET"
+            | b"ROLE"
+            | b"SAVE"
             | b"SCAN"
             | b"SCARD"
+            | b"SCRIPT DEBUG"
+            | b"SCRIPT EXISTS"
+            | b"SCRIPT FLUSH"
+            | b"SCRIPT KILL"
+            | b"SCRIPT LOAD"
+            | b"SCRIPT SHOW"
             | b"SDIFF"
+            | b"SELECT"
+            | b"SHUTDOWN"
             | b"SINTER"
             | b"SINTERCARD"
             | b"SISMEMBER"
             | b"SMEMBERS"
             | b"SMISMEMBER"
+            | b"SLOWLOG GET"
+            | b"SLOWLOG HELP"
+            | b"SLOWLOG LEN"
+            | b"SLOWLOG RESET"
             | b"SORT_RO"
+            | b"SPUBLISH"
             | b"SRANDMEMBER"
             | b"SSCAN"
+            | b"SSUBSCRIBE"
             | b"STRLEN"
+            | b"SUBSCRIBE"
             | b"SUBSTR"
             | b"SUNION"
+            | b"SUNSUBSCRIBE"
+            | b"TIME"
             | b"TOUCH"
             | b"TTL"
             | b"TYPE"
+            | b"UNSUBSCRIBE"
             | b"XINFO CONSUMERS"
             | b"XINFO GROUPS"
+            | b"XINFO HELP"
             | b"XINFO STREAM"
             | b"XLEN"
             | b"XPENDING"
@@ -139,8 +273,6 @@ pub(crate) fn is_readonly_cmd(cmd: &[u8]) -> bool {
             | b"ZSCAN"
             | b"ZSCORE"
             | b"ZUNION"
-            | b"JSON.GET"
-            | b"JSON.MGET"
     )
 }
